move add_embeddings() to ingest_utils.py
